+++ core/models.py
@@ -45,7 +45,7 @@ from wagtailmedia.models import Media
 from wagtailseo.models import SeoMixin
 
 from core import blocks as core_blocks, cms_panels, mixins, snippet_slugs
-from core.blocks import ColumnsBlock
+from core.blocks import ColumnsBlock, LinksBlock
 from core.case_study_index import delete_cs_index, update_cs_index
 from core.cms_snippets import NonPageContentSEOMixin, NonPageContentSnippetBase
 from core.constants import (
@@ -1407,6 +1407,17 @@ class MicrositePage(cms_panels.MicrositePanels, Page):
                 ),
             ),
             (
+                'links_block',
+                StreamBlock(
+                    [
+                        ('link_block', LinksBlock()),
+                        ('text', RichTextBlock()),
+                    ],
+                    template='microsites/blocks/link.html',
+                    block_counts={'text': {'max_num': 1}, 'link_block': {'max_num': 6}},
+                ),
+            ),
+            (
                 'cta',
                 blocks.StructBlock(
                     [
@@ -1536,7 +1547,7 @@ class HeroSnippet(NonPageContentSnippetBase, NonPageContentSEOMixin):
     slug_options = {
         snippet_slugs.EXPORT_ACADEMY_LISTING_PAGE_HERO: {
             'title': 'Hero for the Export Academy listing page',
-            'page_path': ('/export_academy/upcoming-events/'),
+            'page_path': '/export_academy/events/',
         },
     }
     title = models.CharField(